pyproject.toml
[tool.poetry]
name = "247csa"
version = "0.1.0"
description = "247CSA Project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"  
django = "^5.0.2"
celery = "^5.3.6"
redis = "^5.0.1"
mysqlclient = "^2.2.1"
gunicorn = "^21.2.0"
flower = "^2.0.1"
django-environ = "^0.11.2"
social-auth-app-django = "^5.4.2"
pillow = "^11.1.0"
google-auth = "^2.37.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.157.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
openai = "^1.59.4"
django-crm = "^0.9.0"
arrow = "^1.3.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.52"
django-compressor = "^4.5.1"
django-simple-pagination = "^1.3"
psycopg2-binary = "^2.9.10"
whitenoise = "^6.8.2"
django-storages = "^1.14.4"
djangorestframework-simplejwt = "^5.4.0"
firebase-admin = "^6.8.0"
drf-nested-routers = "^0.94.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 